pyproject.toml
[project]
name = "gmailservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.129.0",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "alembic>=1.13.1",
]
